refactor tag mgmt
